git-svn-id: svn://127.0.0.1/flaskdump/trunk@3352 386f392f-2c41-1343-9bf7-4b6fedb51b34

+++ liferay-plugins-sdk-6.2/portlets/flask-user-tailgate-portlet/docroot/WEB-INF/service/com/rumbasolutions/flask/service/TailgateInfoService.java
@@ -103,7 +103,7 @@ public interface TailgateInfoService extends BaseService, InvokableService {
 
 	@Transactional(propagation = Propagation.SUPPORTS, readOnly = true)
 	public com.rumbasolutions.flask.model.TailgateInfo getTailgate(
-		int tailgateId);
+		long tailgateId);
 
 	public com.rumbasolutions.flask.model.TailgateInfo copyTailgate(
 		int tailgateId, com.liferay.portal.service.ServiceContext serviceContext);
