+++ liferay-plugins-sdk-6.2/portlets/flask-user-tailgate-portlet/docroot/WEB-INF/service/com/rumbasolutions/flask/service/TailgateInfoServiceUtil.java
@@ -113,7 +113,7 @@ public class TailgateInfoServiceUtil {
 	}
 
 	public static com.rumbasolutions.flask.model.TailgateInfo getTailgate(
-		int tailgateId) {
+		long tailgateId) {
 		return getService().getTailgate(tailgateId);
 	}
